Display 'deleted' rather than 'deletted'
The regexp used to display an event in the present tense was incorrectly
turning 'deleted' into 'delet', which resulted in it showing up on the
site as 'deletted'.  Rather than try to improve the regexp, I've decided
to hardcode the present tense versions of the states we're expecting.

There are other states in the versions table that I'm not dealing with.
These are 'archived', 'draft' and 'scheduled'.  This is because
'archived' is already shown as 'editioned', 'draft' is already in the
present tense and I'm not sure whether 'scheduled' should show up or
not.  If we decide that 'scheduled' should show up then we might need to
amend this method.