Force WAP requests to return HTML.
This is pretty nasty, but it's felt that it is better if requests
for WAP return something rather than a 406 error.  By adding the
WAP content type as a synonym of html, all wap requests will be
regarded as html.  This is apparently common behaviour across the
web, and recommended by google mobile search.