Define all callbacks the same way.
We thought it was confusing that the observer notification callbacks
were being defined using `set_callback` whereas the callback to
`#archive_previous_editions` was using `after_publish`.

We also thought it was clearer to combine the two "after publish"
callbacks into a single one given that they are defined in the same
module.

We believe that the different styles arose from an earlier refactoring
[1].

[1]
https://github.com/alphagov/whitehall/commit/1d25672d76b3c8c6b488d203132edf88f8496d80#commitcomment-2008315