Revert "Allow deletion of documents published in error."
This reverts commit 1add51d29932194f198e732de9563529b105c27d.

I've ended up manually reverting the commit because of the amount that's
changed in the last 8 months!

We no longer need the ability to delete published documents, as it's now
possible for GDS editors to unpublish published documents, which has the
same effect.

One thing to note is that I've removed the ability to delete a scheduled
document.  It wasn't entirely obvious that this was previously possible,
but the presence of `only_edition?` in `Edition#deletable?` made it so,
providing there was only a single edition of the document that had been
scheduled.  Given that you can unschedule a scheduled document, I'm
happy to remove the ability to delete them.

I've also removed the ability for an archived document to be transition
to the deleted state.  It's not clear to me why that was added in the
now-reverted commit and, as I can't see an obvious use for it, I've
chosen to remove it.