Display embedded youtube videos on policies
I was investigating the `undefined|0|TypeError: 'null' is not an object`
error that we see when running cucumber features, when I realised that
we weren't displaying embedded youtube videos on policy pages.

The problem was caused by commit
ef0e068d5eb70c493aadfe53db6bb11fa0603593, which fixed the bug whereby
videos weren't showing on speeches, but inadvertently broke them on
policies.  The problem was compounded by the accidental removal of the
cucumber assertion in a1a4d168f713b60db337d85223b32b634288f71a.

I've fixed the problem for policies by adding the `document` class to the
div containing the `body` class in the `policies/_document_content`
partial.  This combination of classes is what
`application/on_ready.js` looks for when deciding which elements should
render embedded youtube videos.

*NOTE* Although I can't see any obvious visual side-effects of adding
this additional `document` class, it's very possible that there are/will
be.

*NOTE* It feels as though a more robust fix would be to add a specific
class to the containers whose contents should contain embedded videos,
but I'll leave that for another time.