Adding explicit date for consultation response
This commit adds the ability to choose a published date for the
consultation response. Previously the system was relying on the
attachment creation date for the published date, which will break
imports of old consultations as well as being too much inference.

The new published date does not affect the visibility of the response -
it's only descriptive. This means that if the date is in the future,
a response attached to a published consultation edition will still be
visible to the public.

I don't particularly like the way the the published date pollutes the
attachment partial, but after chatting with Tom we agreed this needed an
overhaul soon anyway and following the existing local variable approach
was acceptable for now.

Also included bonus cucumber scenario for adding a response to a
consultation.