Added data migration to upload DFT consultations.
* This does not include the attachments which will be attached separately.

* There were 3 consultations with closing dates *before* their opening
dates. I have corrected these within the data migration, because I
couldn't work out an easy way to edit the CSV while keeping it in
correct UTF-8 format.

* I've tried to re-use code from other similar data migrations. Although
I'm not convinced it is as simple as it could be - I thought it was
better to be consistent.

* There is no first publication date in the CSV file, so I've taken it to
be the same as the opening date.

* There is no participation data, although there are (empty) columns in
the CSV file. I've added code to cope with having this data, although I
haven't dealt with the case where a form file needs to be attached.

* I'm not using the "consultation status" column, because the system
derives this from the opening, closing & response dates.

* I'm not using the "old url" column, because it doesn't seem useful
here.