Allow overriding of asset_host in staging
Other apps seem to be loading the initializer and overriding the value
of asset_host correctly, but in whitehall the value set in the
initializer doesn't seem to over-write the value set in the the
production.rb. We couldn't work out why.

Secondly, it's important that the Whitehall.asset_host returns the
correct thing.

This solution is a bit hacky, but was the best I could think of for now.
The initializer should override the ENV var value, and that will be read
later in the initialization process.