Reduce the surface area of the Govspeak Helper API.
In order to make this helper clearer, I'm trying to slim down the
number of public methods. I'll then try and rename methods to make
things clearer.

Obviously this is only a very minor refactoring, and much of this
probably belongs in a class rather than as loose methods, but it's a
good start.