Attachment information can now vary per edition
This commit extracts all the 'file-specific' information from Attachment
and adds it to a new class: AttachmentData. The Attachment class is now
one-per-edition, allowing title updates to work correctly.

I've deliberately left the tests with duplicate factory and attribute
setup in them, and used the refactor to understand the way the tests
work better. There's a lot of work to be done to clean them up, but I
wanted to keep this change to a reasonable size.

Still to do: need to copy the assets over to the attachment_data
location.