Ensure all attachment join models have `dependent: :destroy`
This behaviour was inadvertantly lost in 7885f7adc.

To avoid this happening again, I've encapsulated the behaviour of an
attachable join model into a method. One side effect of this is that
`EditionAttachment`'s `accepts_nested_attributes_for` regains its
`reject_if: :all_blank` option; all tests pass so I am confident
that this isn't an issue.

I've also added a missing test for one of the join models, which
did not previously exist.