Skip attachment on error but upload consultation
If we are unable to fetch an attachment for some reason, we'll still add
the consultation, but skip that attachment. Failures will be logged, and
will need to be manually cleaned up (this was discussed with Pete).

I've also added code to record the source url of an attachment so that
we can in the future do redirections of the attachment urls.