A number of organisations have invalid URLs and cannot be saved.
It looks like the URL validation was added in
bbfe9d299ef36c3e19760bcc524ddec84662d43f some months after it was
possible to set the URL through the user interface in
761bd9e377cfea042e659fa4b04f2c3bf4fad3fe. A number of invalid URLs seem
to have crept in in the meantime.

This data migration makes them all valid, mostly by setting the URL to
nil which is allowed, or in one case setting the intended URL which was
missing the protocol prefix "http://" in the production data.