Set most organisations status to "joining" i.e. "Coming soon".
We want most organisations to be labelled as "Coming soon" ready for
when the 1st departments go live.

Re-indexing of ministerial roles normally happens after every save of an
organisation. I used the `.while_disabled` method I added earlier to
avoid this, but I added a single re-index at the end just to be sure.