Add a generator for data migrations.
I tend to use the built-in Rails generator when I create migrations,
at least in part, because it avoids typos in the timestamp in the
filename like the one that was fixed in [1].

This adds a simple custom generator for data migrations. It just creates
a blank file in the `db/data_migration` directory with a timestamp
prefix in the filename.

[1] 714463e832dc94b61f5a8fed5fa44c8092f866ad