Run data migrations in timestamp order.
Since data migration files are organised and named in a similar way to
standard Rails migrations, I assumed they would be run in timestamp
order. This caused a problem when two data migrations were run within
a single deployment in the opposite order I had expected.

To avoid anyone else having the same problem, I've changed the behaviour
so that data migrations are now run in timestamp order, with the
standard proviso that they have not been run already.