Set most organisations status to "joining" i.e. "Coming soon" (again).
The previous attempt to do this in [1] failed, because there was an
invalid organisation which had no alternative format provider contact
email, even though an edition was using it as its alternative format
provider.

I've added a new data migration (20121024102418) which will fix the
problem above by unsetting the alternative format provider on the
relevant editions. Although this will lose data, I believe only one
edition & organisation is affected on preview and it's unlikely
there are many more on production.

Two other recently added data migrations failed when run on preview
due to the problem fixed in the above data migration. I have
therefore changed their timestamp prefixes so they will run after
the fix (now that data migrations are guaranteed to run in timestamp
order). It should be safe to do this, because the data migrations
that failed have not been recorded in the system and will be run again
anyway.

[1] 76d3d12f4288a3da444046c12c41d9b04b3068f3