Remove Google maps
Google maps were only being used in one place. They used to be in
govspeak but that functionality was removed. The only place left was on
organisation pages. However, the styles were such that they weren't
actually displayed there.

This should remove google maps and all traces to them from the project.