Define atom feed route differently to avoid problem in controller tests.
I want to change the `match '/home'...` statement in `routes.rb` to
redirect to another path, but doing this meant I ended up with the
following error in the `HomeControllerTest` tests which test the atom
flavour of the action :-

    ActionController::RoutingError: No route matches {:format=>:atom,
:controller=>"home", :action=>"show"}

I never really got to the bottom of it, but this change keeps the output
of `rake routes` the same and seems to have made no significant change
to the external behaviour, but with the added advantage that I can make
the change I want without the above error occurring.

Thanks to @chrisroos for the suggestion of using the `:constraints`
option.