Allow writers to create Statistical data sets
Writers can now create Statistical data sets in the admin interface.  A
Statistical data set may have the following:

* Title
* Summary. Currently displayed on the Statistical data set page shown to
  the public.
* Body
* Any number of attachments.  Having chatted to Pete, these are rendered
  in the page through the use of the markdown helpers. *NOTE* We've
  enabled the attachments to have all the reference fields (ISBN, Unique
  reference, Command paper number, Order URL and Price in £s) as that was
  the easiest way to allow an attachment to have a Unique reference, which
  they will need.
* An association to a document series that this Statistical data set is
  part of.  This is currently the only way to make a data set
  discoverable: If it's not part of a document series then it won't be
  possible to navigate to it from the frontend.
* Any number of producing organisations, although in reality we only
  expect there to be a single producing organisation.
* The email address from where the attachments can be ordered in a
  different format.

Most of the functionality in building this 'came for free' as we're
already doing it for lots of other document types.  This made it quite
hard to ensure that I'm testing everything as I should be, so apologies
in advance if we find somethings that aren't tested, or aren't tested as
thoroughly as they could be.