Remove old home page and redirect to new home page.
This introduced a slight complication with the cucumber tests, because
many of them were using the global nav to access the public-facing pages
of the site. However the new home page doesn't yet have this global nav,
because it would only encourage people to attempt to access pages which
are behind HTTP auth.

So for the moment, I've substituted the organisations page for the
home page within the cucumber tests. This seems to have worked
well in the sense that I only needed to make a few small changes to the
cucumber tests.

One slight hiccup was that the app now seems to need an
OrganisationType "Ministerial department" in order to show the
organisations page without an error. So I've added something to the
cucumber environment to ensure that this is present.

It would be nice to move the atom feed into an action with a better
name, but this can be done separately.

It would also be nice to move the new home page from the sunset action
to a new home action, but again this can be tackled separately.