Start building the PublicationUploader
There's more work to do on this uploader.  I've been copying the tests
from ConsultationUploader and amending as necessary.  I've also been
adding little helper objects to handle the transformation of a value
from the spreadsheet to an object that's required by the Publication.
The next test to copy from ConsultationUploader is 'alternative format
provider is set from organisation'.

I've introduced `Person#ministerial_roles_at` to help turn a person
(identified by a slug in the CSV file) into the role/s they held at the
time the publication was published.  The method is tested but I wouldn't
be surprised if there are some edge cases that appear...

I've introduced a map between the publication type 'slugs' in the CSV
and the actual Publication Types in the system.  This is because our
Publication Types are faux ActiveRecord objects that don't provide a
mechanism for finding them by arbitrary atributes.

We'll probably want to add some error handling around most of the
`PublicationUploader#upload` method to ensure that the import process
continues even if we hit something unexpected.

*NOTE* I'm currently passing a logger and line number into each of the
little helper objects.  This provides better log output (e.g. 'Row 2:
Document not found') but it's a bit dirty.  It might be better to pass
an errors object (or just an array) into the helper objects and then
have the Uploader print the line number and error messages.