Include LimitedAccess mixin in Edition
...rather than in Publication.

We'll need to define an ActiveRecord 'scope' to select objects
accessible to a particular user, and it feels like that belongs in the
LimitedAccess mixin. This method will need to be callable on all edition
types, hence this refactoring.

The behaviour is enabled/disabled according to the can_limit_access?
method.

feels reasonable that this would
be in the LimitedAccess