Ensure downloaded files are actually available.
The block form of `Dir.mktmpdir` removes its contents after the
block exits, so we cannot easily use it will retaining the interface
of an attributes hash for `PublicationRow`. To work around this,
I've had to add a `cleanup` method to allow the row to store data
in a temporary directory, but not leave files lying around once that
row has been successfully imported.