Make test names less ambiguous and easier to read.
I found the phrase "admin preview" to be very confusing given the
existence of a preview environment and its relevance in some tests.

The distinction between "preview" and "public preview" was lost on me
initially. I believe the intention of these tests is to distinguish
between the internal and public hostnames for a given environment.

I've also re-worded the english in the test names a little with the
intention of making them more readable.