Create resized images for use in other parts of the site.
Due to the helpful methods that carrierwave tries to create for
accessing versions, we can't call the versions '960' (since '960'
isn't a valid method name). Technically we could patch carrierwave
to fix this, but it would be ultimately so awkward to actually
use the created methods (image.file.send(:960)) that it didn't feel
worth it.