Use speech as a representative concrete edition subclass in tests.
This seems to be the style that other tests in this file have taken and
I can see the advantages of using a real edition subclass instead of a
generic edition for which route generation is non-standard. So I've
decided to bring the two tests I modified in [1] into line with the
other tests and use a speech in each case.

[1] 898991205d23a79c07a127a03bdedcca661add0d