Save CSV rows which caused errors for further processing.
The idea here is that any row which cannot be properly processed
should be output into another CSV file, along with the error, to be
passed back through the (non-ideal but currently unavoidable) human
cleanup pipeline. This file can then be cleaned and sent for
importing again, until no errors are returned.