Link publications to data set documents not editions
It's important that publications point to the document itself, not to
editions, so that new editions of data sets are still linked correctly.

I've followed the pattern of related policies: there's lots of similar
code, but I've not yet merged the concepts as it's not clear what the
best abstraction is yet.

The included migration will remove any associations created, but they
have only existed since yesterday and haven't been pushed to production
yet, so we're safe.