Require every document to have at least one organisation
The validation itself is fairly trivial, but widespread changes were
required throughout the functional tests and also some cucumber tests.

Speeches are a special case because they have a callback which sets the
organisation from the role appointment. I changed the timing of the
callback to before_validation (instead of before_save) so that the
organisations field will always be set properly before the validations
run.

The tests needed to be modified to always set an producing
organisation.

In the functional tests I introduced a new method
controller_attributes_for_instance to generate controller attributes
from an object instance, including the associated organisations. This is
for use doing put :update to modify an existing record.

I also had to modify some factories to generate the organisation
association using an after :build callback.

Fuctional tests which disable database access and use pure stubbing
(such as Api::DetaildGuidePresenterTest) needed to be changed to
explicitly stub the associated organisations as well, since the
stub_record and stub_edition helpers will otherwise trigger database
accesses during the after :build callbacks in the factory.