Don't import publications if attachments are missing.
Previously, attachments that couldn't be downloaded, or which had
no titles or URL, would've been silently skipped.

I had to remove some tests and 'helpful' behaviour within the
`AttachmentDownloader` class, so that it would knowingly create
invalid attachments. This way we rely on the validations to catch
issues with the data, which is probably sensible anyway.