Make re-writing of admin links more robust.
The previous logic for re-writing admin links was very relaxed and in
particular it did not scope the supporting page lookup to the edition
included in the path.

Since there is no restriction on the uniquness of a supporting page
title, multiple supporting pages could exist with the same title, but on
different editions. This led to a bug where the wrong supporting page
was being found just because it was found first in preference to the
correct one.

This commit explicitly addresses the latter problem by scoping the
lookup of a supporting page to its parent edition.

However, it also changes the logic so that we only ever re-write links
to editions and supporting pages. Previously it is conceivable that an
admin link to e.g. an organisation might have been re-written as a link
to an edition if they happened to have the same ID.