Introduce a new consultation uploader for DCLG consultations.
This new uploader is based on the same pattern as the latest
publication uploader, using a row class to gather the attributes
from each CSV row required to build a consultation.

The actual DCLG data will be included in a later commit, once some
final checks have been made to ensure it imports correctly.

The previous uploader hasn't yet been replaced.  This will be done
on the next commit.