New upload for the DCLG publications that previously failed
The previous DCLG publication upload
(20121029160239_upload_dclg_publications.rb and corresponding csv file)
stopped at row 948 because of a network error.  I've created a new file
containing rows 949 to the end of file.  I've run it locally against a
recent dump of the production database and the rows all appear to import
correctly.