Ignore case of headers in CSV files
The latest DfT FOI publications have a column titled 'urn', as
opposed to 'URN' which is what we'd been received previously. This
commit makes the headers case-insensitive so that we can happily accept
either 'urn' or 'URN'.

I've updated the ConsultationRow and PublicationRow so that they deal
with column headings in lower case. I've also updated the
DclgConsultationImportTest so that I can run the test from within
TextMate, by referring to the absolute path of the sample CSV file.