Fix delivered_on and first_published_at of DfT Speeches
In the original DfT speech uploads
(20121012135700_upload_dft_speeches.rb) we were parsing the date from
the CSV file as a timestamp.  Any date during BST would be converted to
a timestamp at midnight (BST) on the same day, which was then converted
to 23:00 the day before (UTC) when stored as a date.  The effect was
that speeches were appearing to be delivered the day before the date we
had in the CSV file.

Additionally, we weren't setting the first_published_at attribute which
resulted in it being set to the datetime that it was published in our
system. To avoid any ambiguity, I've set first_published_at to 13:00
(rather than allowing it to default to midnight) on the same date as we
now have for delivered_on.