Fix delivered_on for round 1 of DCLG Speeches
In the original DCLG speech uploads
(20121031105330_upload_dclg_speeches.rb) we were parsing the date from
the CSV file as a timestamp.  Any date during BST would be converted to
a timestamp at midnight (BST) on the same day, which was then converted
to 23:00 the day before (UTC) when stored as a date.  The effect was
that speeches were appearing to be delivered the day before the date we
had in the CSV file.

Additionally, I've chosen to update the first_published_at attribute to
13:00 (rather than allowing it to default to midnight) to avoid any
ambiguity between the delivered_on date and first_published_at datetime.