Publication upload had incorrect json column name.
This was my fault.  I first tested against an earlier version of the
CSV file with the correct column headers.  Then, after replacing
with the latest version, I ran the data migrations and checked the
output.  I only did a quick visual check, and mistook attachments
embedded in the page copy for the actual attachments.

The migration hasn't been run on production, as Chris Parsons found
this discrepancy whilst doing the staging release.