Import DCLG consultations, round 2.
Seems to work without problems locally. I had to modify the file to
handle all the empty 'columns' (CSV freaks out if there are empty
headers and you're using a header converter), and to ignore the
erroneous dates at the end of the file. Robin confirmed that this
data isn't needed, but felt it was more expedient to modify the file
here rather than generate a new one. He is fixing his source data to
remove the extra columns.