Handle CSV file with UTF-8 Byte Order Mark (BOM) at start.
@davidh's character_encoding_cleaner adds a UTF-8 BOM to the beginning
of a file in order to encourage apps like MS Excel to open it with the
correct encoding. Unfortunately the current uploader logic cannot handle
this BOM which results in the "old_url" column heading not being found,
which in turn means that a blank "old_url" is used to determine whether
or not the row has already been imported. The first one may be imported
with the blank value (which is undesirable in and of itself), but
subsequent rows will not be imported at all, because the blank source
url will then already exist.

This commit demonstrates how to handle the BOM correctly. Although it's
not actually necessary in this particular case, because I have already
manually removed the BOM from the corresponding CSV file in the repo.
However, this approach works whether or not a BOM is present, so is
probably a good way to go in the future.