Only show featured dcoument on org page if image has been virus checked.
This is to avoid displaying the placeholder image in such a prominent
position. See https://www.pivotaltracker.com/story/show/38484895.

We use the fact that the file path for the image is the path where the
image will be placed once it has been removed from the virus scanning
quarantine to determine whether the image is ready to display.