Show the 10 most recent documents in the home page atom feed.
* There was a flaw in the controller action and in the test so that
previously we were showing the 10 *oldest* documents by "published_at".

* Using `Edition#in_reverse_chronological_order` and
`Edition#timestamp_for_sorting` brings this feed into line with most of
the rest of the site. This sorts by `Edition#first_published_at` for
most document types, but `Publication#publication_date` and
`Speech#delivered_on` as special cases.

* Using `Edition#timestamp_for_sorting` for the atom feed entry
"published" attribute should mean that we only show documents that are
genuinely recent i.e. not just ones that have been *added* to the system
recently. This is important given that we will be adding "legacy"
documents for new departments in batches in the future.

* It seems most consistent to use the latest `timestamp_for_sorting` for
the overall feed "updated" attribute. And using `Edition#published_at`
for each entry's "updated" at seems as good as anything.

* See https://www.pivotaltracker.com/story/show/38882917 for further
details.