Fix non-deterministic test.
This test was not deterministic even before I made the change in
f6db51f39a737f072ee0c3aacf0909be781b3425. I saw it fail locally, but not
repeatably before making that commit and tried to fix it. Unfortunately
I rushed to an assumption that the publications are sorted by
publication_date when in fact in this case they are sorted by
published_at (incidentally this is something that should probably
change). I have now changed the test to be truely deterministic by
setting values of published_at. Something that's very confusing is that
time is frozen by Timecop.