Revert "Only show featured dcoument on org page if image has been virus checked."
This reverts commit c4b2aea18cbe55ca40ae4fc17f07fe1746b3b410.

I'm reverting this, because the frontend server doesn't appear to have
access to the uploaded or (more importantly) the scanned files which
seem to reside on the backend server. This means that the edition
organisation image is never reported as ready and so no featured
editions are ever displayed on the org page.