Show 10 most recent publications on feed only
* Using `Edition#timestamp_for_sorting` for the atom feed entry
"published" attribute should mean that we only show documents that are
genuinely recent i.e. not just ones that have been *added* to the system
recently. This is important given that we will be adding "legacy"
documents for new departments in batches in the future.

* It seems most consistent to use the latest `timestamp_for_sorting` for
the overall feed "updated" attribute. And using `Edition#published_at`
for each entry's "updated" attribute seems as good as anything.

See https://www.pivotaltracker.com/story/show/38882917 for reasoning.