Always allow publication date to be up to 5 years from now.
Presently it's set to 5 years from whatever the publication date
currently is, so editing a publication with the date set to 2006
would only then allow updating it upto 2011.

An interim way to cope would be to set it to 2011, save, and then
set it to a higher year, but this commit will always set the end
year to 5 years from now.