Reimport some publication errors.
I had to do a bit of work to the attachment cache to get the
majority of these files to upload, but most of them go in now.

It's worth noting that most of the second file, however, does not,
because the attachment URLs don't seem to point to actual files, but
index (HTML) pages.

As an aside, I'm not sure that the validation error messages are coming through when an zip file contains an attachment of an
invalid type; I *think* it should explain what the invalid type was,
but I'm just seeing that the top-level ZIP is invalid, which isn't
as useful at all.