Script to fix missing attachment data
When running a migration on the 24th October, we corrupted the live
attachment data. This migration fixes the data by recreating the attachment
join rows for each attachable type.

We will need to dedup the attachments that editors have re-uploaded
after noticing the data error: see this story:

https://www.pivotaltracker.com/story/show/39563845

The sql represents the old uncorrupted data: this will need to be loaded
manually into the production tables before running the included rake
task.