Allow access to statistical data sets to be limited.
This is the same behaviour as is applied to statistics publications.

The default behaviour for statistics publications seems to be to
restrict access i.e. the checkbox is checked by default, but it's not
clear to me that that was the intention. However, after discussion with
Pete, I've made the behaviour for statistical data sets the same.

The latter made things a bit awkward in the controller test, because by
default all the tests were creating editions that were inaccessible. I
got round this by defining
`StatisticalDataSet#access_limited_by_default?` and stubbing this in the
setup to return `false`. And stubbing it to return `true` for the
relevant tests.