Existing datasets with no access limiting should not limit access.
In [1] I used the `after_initialize` callback to set the default value
for `StatisticalDataSet#access_limited` to be `true` based on a
conversation with Pete. This was so that new datasets have their "access
limited" checkbox checked by default.

However, I didn't notice that using `after_initialize` would also kick
in for existing persisted datasets and this would mean they would all
appear as if they had access limited, even though they did not, or at
least not until someone tried to edit them.

This commit adds a check for `new_record?` so that access is only set to
limited for *new* datasets.

[1] 1554e5f6717e46fd46324388516a23c86f1d0a28