Implement just-in-time removal of extra quotes
This does require an extra processing step when processing govspeak, but
we decided that it was a nicer user experience to auto-remove these
quotes for the content creators. It also means no manual re-running of
a quote removal step on import.

The implementation steps through each line in the body, so it isn't
wonderfully efficient: I attempted a few other regular expressions and
am open to comments as to how it could be improved.