Add minor and major published versions to editions
We need to be able to tell version number of an edition so we can
display different metadata for some edition versions. The sort behaviour
changes when new major versions are published.

We can also use this information to find the latest published version
and to reliably tell whether we are on the first edition.