Revert commits about re-writing admin organisation links in govspeak.
Two commits [1,2] attempted to allow editors to paste admin organisation
links into document markdown and have them re-written as public
organisation links.

Having discussed this with Pete, we decided that since this link
re-writing isn't complete i.e. it is not done for any other entities
(e.g. roles, people, document series, etc) and it doesn't currently work
for all organisation-related links (e.g. organisation corporate
information pages), it would be better to revert this functionality and
make editors aware that they should use public links for entities other
than documents.

It is not a complete reversion of those commits in that I've chosen to
keep some of the test changes and in particular the tests to check that
pasting in the admin links do not cause an exception.

[1] 2a236d623e391f8e927dd0c767a2e65cb2bdd80e
[2] 8a5b764d6e32b7c7a2f89af1400859d0a44c6841