Re-instate 'Only applies to' on detailed guides.
I suspect this went away because the helper is used in more than one
place (i.e. policies as well as detailed guides), and the copy was
no longer wanted in one of those other places.

I've changed this to return a `span` rather than a `p` tag, because
it's no longer being used as a paragraph, given the loss of the
surrounding text.

I also suspect that `_document_metadata.html.erb` is no longer used;
I will remove it if I can.