Added user index and removed editing of own record
There is now a new "Users" menu option which will show a list of all
users in the system, along with role and email address.

As part of this change we have removed the ability for users to edit
their own records. This includes their email address and their
organisation, which provides the security we need against embargoed
attachments. GDS editors can edit both organisation and email address.