Don't block attachment thumbnails (for the time being).
Blocking attachment thumbnails prevents them appearing at all on
the admin server.  This is because images (including these
thumbnails) are served through the assets host, so don't include the
cookie that would give permission to view the thumbnail.  I am going
to continue to think of a better solution but for now it is better
to show these thumbnails (as long as the attachments themselves
are still blocked).