Skip the assets host for admin requests to uploads.
By overriding path_to_image we can force rails to skip the asset
host in certain circumstances.  We want to skip it when requesting
uploaded files as an admin user, as we need these requests to go
through the attachments controller (which cannot authenticate
admin users if the request comes through the asset host).