Make header validations correspond to what actually gets imported
Some of the fields we required before were incorrect or not actually
used by the row importers. Now the required fields reflect what is
actually imported (with the exception of PublicationRow['price'], which
I'll fix in another commit).

I've also made the validations more flexible so that some fields are
optional.