Don't include sub-organisations in the org count on the index.
I've introduced a more meaningful `agencies_and_public_bodies`
method which better captures the 'political' difference between
those and other modelled organisations.

I am aware that I'm slowly nudging up the number of queries used on
the page, and slowly increasing the limit in the `n+1` test, but I
don't think this is a problem yet. It might be good to have a sense
of what the number of queries actually is when we are *not*
defending against N+1 queries...