Only certain users + GDS can edit fatalities
Fatality notice create and field of operation management is hidden from
most users. To get it, you need to be in an organisation with the
'handles_fatalities' flag set, or a GDS editor.

Everyone with publisher access can still see the fatality notices in the
main editions list and view them, but cannot edit. They are able to
create a new edition, but not submit it.