Integration tests should test the new Import object
We removed the old ConsultationUploader PublicationUploader and
StatisticalDataSetUploader in 2969aab4b, but I didn't notice these
integration tests.

I've changed them to test the Import instead of these old uploader
classes.

I noticed that the heading validations on PublicationRow did not allow
json_attachments to be specified, so I've added that.